design: memberlist 조회 뷰 작성, 리팩토링
