+++ src/apis/workspace/workSpaceMember.js
@@ -1,0 +1,4 @@
+import { request } from '../axios/axios';
+
+export const getWorkSpaceMemberList = (workSpaceId) =>
+  request.get(`workspace/${workSpaceId}/members`);
